requirements/local-linux.txt
asgiref==3.4.1
Django==3.2.10
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-model-utils==4.2.0
Pillow==8.4.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
Unipath==1.1